fix(TCOMP-968): extract AbstractActionParameter from ActionParameter
